refactor(starknet_sequencer_metrics): move parse metric fn to metrics crate
commit-id:c8fd2d12
