+++ crates/starknet_batcher/Cargo.toml
@@ -43,3 +43,4 @@ starknet_api = { workspace = true, features = ["testing"] }
 starknet_infra_utils.workspace = true
 starknet_l1_provider_types = { workspace = true, features = ["testing"] }
 starknet_mempool_types = { workspace = true, features = ["testing"] }
+starknet_sequencer_metrics.workspace = true
