+++ crates/starknet_http_server/Cargo.toml
@@ -24,6 +24,7 @@ starknet_api.workspace = true
 starknet_gateway_types = { workspace = true, features = ["testing"] }
 starknet_infra_utils.workspace = true
 starknet_sequencer_infra.workspace = true
+starknet_sequencer_metrics.workspace = true
 thiserror.workspace = true
 tokio = { workspace = true, features = ["rt"] }
 tracing.workspace = true
