+++ crates/starknet_infra_utils/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod command;
-pub mod metrics;
 pub mod path;
 pub mod run_until;
 pub mod tasks;
